support other var types
